Set Favicon - Round 2
The previous flag Icon was too rounded ... this new one looks more
similar to the one from Apple Emojis, which is what I was going for.
I couldn't just use the Emoji text itself, though, because that is
interpreted however any browser chooses. So when I lookd at it on Mac
it looks great, but when I uploaded it to
https://realfavicongenerator.net/ it came back with a "W" on top of it
(?). So anyway, I found this other flag vector shape in Pixelmator. It's
for the "Golf flag", I just blew it up really big and only used the
flag part. This makes for a simple vector shape that's close enough to
what I was going for in the Apple Flag Emoji, while still being quite
small in file size, convertible on Favicon Generator, etc.

Also, fix my past blunder with not having a /manifest.json and, instead,
having been duped by the app/views/pwa/manifest.json.erb file, which
looks quite similar to what we need for the favicon manifest. (I'm still
not sure what the PWA one is about, entirely.)