Fix bug in "Previous Op" link/logic from Home page
We recently changed the visibility logic for the page nav elements on
the Home Page, but didn't update the "Previous Op" link/logic to match.
Basically, now that we show the "Previous Op" link even when there is no
Current Game, `Game.second_to_last` is incorrect. Really, we just want
to link to the last "game over" Game.