Move `current_user`: Application::Layout -> ApplicationController
While looking into Rollbar integration, I saw that they expect a
`current_user` method on ApplicationController. This just highlights how
ingrained this concept is for idiomatic Rails. Having been so used to it
myself, I've been wanting to reverse my earlier decision to move this
into Application::Layout as well. So this excuse is adequate for me! It
also simplifies things which is a positive 👍