Record Game Score for victorious Games
Game Score is just the duration in seconds, as per typical Minesweeper
rules (lower score is better, and Score maxes out at 999). This is very
simplistic, and there are other ways to calculate efficiency, etc. but
having a solid "score" value is a good start.

We could just calculate Game Score every time it's needed, but I chose
to cache it into the games table as `games.score`. This makes it easily
rankable/filterable later.

Game Score is displayed on the Index page as well as the Game Show page,
after a Game has ended (in victory).

Also, fix a typo found on the Game Index page `dark:text-neurtal-400`...
which turned out to be a happy accident because I like(d) the way the
text stands out in the dark theme when the typical "neutral-400" isn't
applied. So fixed by changing this to `dark:text-gray-500`.

Also, reverse the order of the Game timer and the mines/flags count, to
match the design of the original Minesweeper game.