Add Calc3BV, a service for determining board difficulty
Calc3BV is a Service Object for calculating the 3BV value (3BV is a.k.a.
Bechtel's Board Benchmark Value) for a given Game Board / {Grid}.

A 3BV value is a measure of board difficulty. It represents the minimum number
of left clicks required to win a Game, based on:
- Each opening requires at least 1 click, and
- Each non-mine, numbered (value) {Cell} that doesn't touch an opening
  requires at least 1 click.

References:
- https://minesweepergame.com/statistics.php
- Example 3BV value: https://minesweeper.online/game/3864516014
  - Note to self: I've created/exported a Pattern based on the above
    game called "3BV Test 1" for testing purposes. Scoring works out.
    Setting up a test case using this is... a lot of work. So I'm
    skipping it for now. We still have a very simple test case setup.

Also:
- Add Grid#columns_count
- Fix logic for finding Patterns by `name` attribute.
- Allow for Coordinates::Console objects to be sorted via
  Coordinates#<=>