Add CellChordTransaction + Display of full Game Stats
I wanted to be able to show Efficiency (the total number of clicks vs
the ideal number of clicks--as defined by the 3BV value)...

To support this, we have to:
- Be able to distinguish Chords / Chording clicks as a single click.
  Previously, we were capturing each of the immediate neighbors revealed
  by chording as separate "Reveal" transactions.
Note: Revealing cells by clicking on a number--with the appropriate
number of flags placed in neighboring cells--is called "Chording".

Given this, we can now display the full stats for each Game. When I say
"full" I mean... all those that are displayed on minesweeper.online.
e.g. see: https://minesweeper.online/game/3871722037
- We could add even more stats in the future... but I think they're all
  derivable from the numbers we now have.