Cache Game stats into games table on Game end
Previously, we were calculating Game stats in real time, via the
Game::Stats View Model. This was fine, until/unless we want to be able
to sort, filter, or average out Game stats values. So, in anticipation
of this, we will now cache these values into the games table on Game
end.