Cache display of User stats
Instead of worrying with touching User records to bust the cache, we
create a cache key that consists of the User record plus the number of
completed Games they've participated in. The latter does use a query to
discover, but that's 1 query in exchange for (currently) 12 others. And,
in the case of a cache miss, this query result will be used in
displaying the User stats anyway.

Also, a fix so that both the cache key and the stats generation only
consider completed Games.