Fix Games::Console#reset issues
This is a console-only thing, but when resetting a Game we must also
properly reset the `started_at` Time. This has always been an issue, but
it's risen up now due to our new duration calculation for Game scores.
Without properly resetting the `started_at` time to a valid DateTime,
we were failing with NilMethodErrors when trying to re-play a reset
Game.