Allow Turbo to Cache the Game timer
Previously, clicking away from a game and then using the Back button on
the browser was causing the Game timer to disappear (from the top-right
Corner of the board). This was caused by having added the
`data-turbo-temporary` directive to this element, which prevents Turbo
from caching this value. I had done this because it seemed appropriate
to not cache the value since it is dynamic, but this use case presents
a compelling enough reason to not do this.