Fix cookies.js -- cookie deletion / expiry settings
I had previously made changes to the options argument structure
regarding cookie expiration, but failed to realize that cookie deletion
needed to be updated to match as well. So this commit fixes that, so
that `cookies.delete("name")` actually works now. We just didn't have
anything that was using the code before.

Also, fix documentation to use the new `expiresIn` argument naming, vs
the old: `expires`.