Tidy: Don't prefix View Model method names w/ `display_`
This is an easy trap for me to fall into, but it feels wrong from the
POV of the view template. The message we want to be sending to the
View Model is the simple one: e.g. `best_bbbv`, not `display_best_bbbv`.
Then, the View Model just has to handle its internal needs a little
differently. In this case, we can use the same name, just prefixed with
an underscore to differentiate private methods / intermediate steps.