Show past Games on User Show page
This is a view of all/only Games that the given User has participated
in. And the same is true of the Previous/Next Op arrow links that are
displayed form this view as well.

Of note: Our View Models make it pretty easy to handle this, with no
additional logic added to the views, and no conditionals needed in the
models. We simply add some specialization (through inheritance) on
some of the existing View Models. I'm comfortable with using inheritance
here because these really are one-for-one "is a" relationships, where
(strictly) specialization is needed.