Fix: Scan heroicons when building tailwind.css
I think I've finally found the reason why I've long thought the `size-*`
utilities didn't really work in tailwindcss. I think it's because of 2
things:
1. Not having been scanning SVG files for tailwindcss classes, and
2. Order of precedence when using classes in SVGs is unexpected...
  - If the SVG has `size-6` and the HTMl specified `size-5`, the
    `size-5` doesn't outrank the `size-6`... not sure why.

Number 1 ended up being the reason for a pretty insidious UI bug that I
started running into after changing the CSS in the new `_dev_flags`
partial (previously `_breakpoint_indicator`). What was happening was:
1. I removed the `w-6 h-6` classes from the `_dev_flags` partial.
2. A few other commits were made.
3. Then I started noticing that the "Rules of Engagement" button was
  too skinny...
4. More time passed as I was in the middle of other code.
5. Later, I noticed that the theme selector icon was missing. (These
  2 things ended up being connected, though that wasn't clear at all
  to start with either.)

It took a lot to figure out the bug since simply reverting to earlier
code commits didn't expose it. This is largely because there is some
measure of lag between build outputs for the tailwindcss CSS
file/output. It has to detect a change in order to re-build. Which it
was doing eventually but not in sync with commit changes... as it had
to find the `w-6` or `h-6` class somewhere else first (more later).
So no amount of git-bisect was exposing exactly what introduced the
issue in any understandable fashion.

The answer: The tailwindcss classes in the `class=""` attribute of the
hericons/SVGs were not being parsed by Tailwind. At all/ever. Because it
had never been told to go look there. So, any classes added to these
files would work if, and only if, they were also added to some other
element somewhere else in the project.
- Well, when I unknowingly removed the only instance of `w-6` and `h-6`
  from the project (outside of the hericons/*.svg files), and once the
  twittercss build script next fired, we'd lose the width/heigh
  information for those SVGs entirely.

Apparently, when devoid of width/height specifications, SVGs do weird
things. Sometimes they disappear. Sometimes they ... mess up the layout?
(Not sure, really.)

Anyway, now that tailwind.config.js has been instructed to scan
`./app/assets/images/heroicons/*.svg`, we can start to rely on the CSS
classes added to these files, for the first time.
- In the spirit of this, I've also gone through and removed the paranoid
  `width`/`height`/`min-w`/`max-w`/etc classes from the various
  `inline_svg_tag` calls. Any consistent styles were moved into the
  heroicons/*.svg files, themselves, so that we can just expect the
  defaults throughout the site without having to redefine them all the
  time.
  - If this still doesn't end up working... can add the classes (or
    some portion of them) back later.

Also:
- Add `size-5.5` as a custom tailwind config option. This is useful for
  the size needed by the `x-mark.svg` heroicon. This is better than
  using `size-[22px]` for something that feels like an official, and
  oft-enough-used, feature of the site.