Move "About" page to its own controller
I missed this one before... This was the last controller/view that was
still stuck in the HomeController, when I was wanting them to all be
their own things now.

Also: Remove the `<abbr>` element/tag from `MMS` in the page text since
it was already obvious what this acronym means. The surrounding context
is literally the definition of `MMS`.