Greatly simplify our CSS tooltips implementation
We really don't need the up/left/down/right options... we just need
the "up-right" style tooltips. Accepting this greatly reduces the CSS
code required, which is worth it. It also allows us to simplify the
HTML attributes needed to signal a tooltip in the UI.

Also, add examples for tooltips to the UI Portal