Fix New/Custom Game buttons/form jumping on validation error
The CSS spacing logic wasn't quite right... and was ab it more complex
than needed. Fresh perspective on this helps, as I remember it being
quite hard to follow when I was in the thick of it originally. Anyway,
this commit removes some cruft and fixes the jumping form on validation
error.