Optimize Theme Selector menu CSS; Add `light:` variant
I first worked on this a long time ago now, so my TailwindCSS-fu wasn't
nearly as strong. Also, my HTML/CSS style guide has updated a bit. So
this is an optimization pass. Functionality is mostly the same, but has
been tweaked. I found a few things like:
- Slightly off-vertical-center button
- Barely visible bg color highlighting in light mode
- Slightly too small of a gap-x between icon and text
- The menu was slightly too wide feeling for the content
- Menu button border was only barely visible in dark mode

In the spirit of optimization, I've also added a `light:` variant to our
TailwindCSS config.

```
 # Instead of:
 hidden dark:inline

 # We can now do:
 light:hidden
```

I don't think this should be overused... I just want to use it to avoid
the scenario of having to disable something in dark mode just so it can
enabled in light mode.