Refactor "Records" Page -> "Metrics"
It was feeling like there wasn't enough semantic difference between
"Archives" (as in Sweep Ops Archives) and "Records". Archives are / can
be Records and vice versa. Metrics makes it more clear, I think, while
also being more encompassing of different kinds of scores, rankings,
etc. (Naming things is hard.)

"Record" is also a bit confusing as a programmer. Especially in a
framework that uses ActiveRecord at its core.

Also, Change the resources for Metrics - Index to Metrics - Show.
Because there isn't a listing of Metrics to display. We're just showing
the set of all Metrics that we display.