Adjust page nav elements in various contexts
1. Don't show Previous Op button on Home - New view

I keep waffling back and forth on this one. Current thinking is that
the Previous Op button is distracting and in the way--especially on
Mobile--when there is no current Game and the War Room is displaying
the New Game view.

2. Show a "War Room" (basically, but not always: Current Game) button
  when viewing past games from the context of the Sweep Ops Archive.

Another one that I've waffled on, but currently:
It feels right to have a right arrow that goes from the Sweep Ops
Archive back to the War Room if there's a left arrow that leaves the
War Room and goes to the Sweep Ops Archive.

3. Don't show the "Close" button in page nav when navigating past Games
  from the context of the Sweep Ops Archive.

If I decide to later adjust the context of the Previous Op browser
function via War Room to be specific to the War Room concept... then the
Close button could be relative to that. i.e. it could close the past
Games browser and return to the War Room. i.e. the URL would need to
look like: `war_room/games/49`.

This commit also does just that when browsing past games via the User
context. So when browsing past Games via a User's "Bests" list, the
Close button will reveal the User again. This is possible, in this case,
because we do scope the Games browser to the User record. i.e. the URL
looks like: `/users/e686d9ba-b345-4eb7-9bf2-890d21f5b454/games/49`.