Fix User -> Game nav context to show Usernamer in title
Previously, there was no difference in page title between a Game show
page and a User -> Game show page. This changes this to show the title
as `<username> -> <Game <game number>`.

As the view template is now in place to handle this, this could
potentially be expanded later to show other User-specific elements
within the User -> Game nav context.