Refactor `css_class[es]` -> `css`
Disambiguate the duality between `css_class` sounding like a singular
thing (when it could be a String that represents multiple CSS classes)
and the need for these values to flexibly represent multiple things. I
don't like taking the "always represent it as multiple" approach. So, I
think just using `css` by itself is a great compromise. I recently
started doing this for e.g. `button_css` and liked it, so this commit
applies this change wholesale.