Simplify <hr> tags CSS
Since all of our <hr> tags use the same styles (still), we can move
the base style up to application.tailwind.css. Local overrides can still
be made, of  course, but this is a good starting point. And it avoids
always having to remember / look up what classes are supposed to be
added to make future <hr> tags consistent.