Remove ActiveModelWrapperBehaviors
This was added early on in the project and ultimately ended up being a
poor fit for this project. The advantage of using `to_model` is that
Rails can treat non-ActiveRecord models as if they are an ActiveRecord
model. This is useful for e.g. passing wrapper objects directly into
`form_with`, etc. But using it as the primary internal reference to a
View Model's wrapped object does add to cognitive load vs just referring
the wrapped object by its type name.

Also, just refactor away from `to_model` methods in general. The
abstraction needs to be worth the cognitive load.