Show completed Games w/in a turbo-frame on Metrics page
Instead of visiting the regular Game - Show page, we can easily just
display the Game on the Metrics page using a turbo-frame tag.

I want to improve this still, but this is a good first step.

To support this, I've broken up the Metrics::Show view model a bit more,
to reflect the page structure as it is now, and as I envision it
unfolding soon (to include top Users, etc.).