Don't show Game Engagement Time Range on Games Index pg
We don't really need the Game Engagement Time Rnage to be shown to
uniquely identify games in the Games Index page anymore. It's a poor
identification mechanism anyway. Better, is the one that's been added
recently, which is starting by visiting a User Show page and then
indexing Games from there. And or by "Bests" via the Metrics page. And
whatever else will be added in the future. These are more practical
approaches than scanning for a logical time range string.