Add "Close" button to game "frame" in Metrics Show page
First, my model for the new Close button was the Banner view code. But,
looking into that, it became clear that the location of the CSS styling
was unnecessarily abstracted, so I started by simplifying that. Then
copied the result into the new button on the Metrics Show page.

Add frame_controller.js to handle show/hide of the game "frame". I'm
using "frame" in quotes to signify the CSS-styled box that holds the
actual <turbo-frame> into which the Game Show view loads.
- On show: we animate the "frame" into existence
- On clicking the new "Close" button, we animate the frame out of
  existence, and then remove the contents of the <turbo-frame> element
  contained within.

Further, when closing the Game "frame", we should also clear the active
link indicator in the Game "Bests" list. We accomplish this by
dispatching a "frame:hide" event and then listening for it on the
window object in the context of active_links_controller.js.
- Dispatch: `this.dispatch("hide")` (See: frame_controller.js#hide)
- Listen: `data-action="frame:hide@window->active-link#clear"`
  (See: the container element for data-controller="active-link")

And one more detail: Instead of dispatching a custom event when clicking
on a Game link, we can just listen for when the <turbo-frame> element
loads, and use that to "show" the frame (i.e. call
frame_controler.js#show).
(See: `action: "turbo:frame-load->frame#show"`)

Of note: We use el-transition to define animation start/end states for
showing/hiding the "frame". This feels better than just a stark
show/hide event. This was modeled after the Collapse control at
app/views/application/_collapse_container.html.erb.