Show nav link underlines outside of header/footer
It's hard to tell when links are clickable if they don't have the usual
underline effect. e.g. with the new Game show/hide toggle links on the
Metrics Show page, but also for the old Sweep Ops Archive filter links
at the top. For Sweep Ops filter links, it doesn't look so bad to have
the underlines if the abbreviation isn't included.