Scroll down to "frame" after clicking "Bests" Game link
This applies to Game Show links on both the User Show page and the
Metrics Show page--where the Game is shown in a turbo frame w/in the
same page.

To support this, we add scroll_to_controller.js, which comes from
another one of my projects (but simplified a bit). Other functions that
could come in useful for scroll_to_controller, but would be purely
future-coding to include right now:

```
jumpTo({ params: { targetId } }) {
  const $target = document.getElementById(targetId)
  this.#doScroll($target)
}

jumpToSelf(event) {
  this.#doScroll(event.currentTarget)
}
```