¡! Restructure Views into Current/Just Ended/Past Games
Previously, I was struggling with maintaining view logic/conditionals.
Particularly, when to render what while keeping in mind the major Game
states/context. The primary Game states are:
- Current Game: Standing By + Sweep In Progress
- Just Ended Game: Within 0.5 seconds of Alliance Wins or Mines Win
- Past Game: > 0.5 seconds of Alliance Wins or Mines Win

In particular, past Games will have different display needs based on
different sub-contexts as well:
- Current Game:
  - Game status ("Standing By" vs "Sweep in Progress")
  - Rules of Engagement
- Just ended Game:
  - Game "Title"
  - Game status ("Mines Win" vs "Alliance Wins")
  - "Signature" section for signing Username + <hr> above it
  - Metrics + Duty Roster
- Sweep Ops Archive: Standard view of past Games.
  - Game "Title"
  - Game status ("Mines Win" vs "Alliance Wins")
  - Metrics + Duty Roster
- Metrics -> Games:
  - Game "Title"
  - Metrics + Duty Roster
- User -> Games:
  - Game "Title"
  - Metrics + Duty Roster

Supporting all of this + thoughts about how these views could diverge
into the future was getting too confusing using a single set of
views + view models within app/views/games. Not only was the directory
growing large, but the view templates were large, getting more complex
despite having View Models backing them, and discrete sections were
getting too hard to maintain and move around as the design changed in
ways that needed to be consistent across more than one Game state.

So, to fix, I:
- Broke view templates out into partials
- Made separate View Models to manage each partial
- Made separate Game state-related sub-directories for view templates +
  View Models
- Pushed the conditionals up the stack as far as they would go so that
  the major game states sub-directories/view templates would be free of
  this level of logic control

Now, when a request comes into the Home page:
- Home#show
  - Renders the page_nav + welcome banner
  - Then appeals to the Home::Show view model to pick a major Game state
    View Model + view templates (this is the "one" conditional)
- Game State view models + templates take it from here.

Of note, there is a good amount of duplication with this approach. But,
actually, most of it ends up being very simple setup code like building
View Model objects from different contexts. This is acceptable
duplication as it is precisely the way in which we can control the
differentiated experiences by context.

There is probably a long list of other improvements to the code made in
this commit, though I tried to keep things as focused on the restructure
as possible.