Allow for horizontal scrolling on Game board as needed
I swear that previously, allowing for this would leave a permanent
horizontal scroll bar on the div in FireFox. This seems to be gone in
the current release, though... and this is a much better experience
anyway. Else, the Game board can appear off screen and the entire page
has to scroll to see it.