Rename _frame_container to _display_case
I feel the latter is a more memorable / unique name for what this is:
An attempt to frame up a past Game so as to highlight it and distinguish
it from the surrounding page.

Also, alleviate the Shotgun Surger code smell created by defining this
turbo_frame_name value all over the place by having just one point of
origin for it: `Games::Past::Show.turbo_frame_name`.

Finally, use a View Model for defining customization within the
_display_case partial vs passing in locals. This allows the
_display_case partial and the "Display Case" concept to be independent
of the context-specific implementation details.