Add "copy" icon to Game Number in title
This allows for easily sharing Game links.

For the animation... I started with TailwindCSS's `ping` animation but
sped it up, decreased the scale, and changed it to a single iteration.
This allows us to listen for the `animationend` event and dehighlight
the icon on that event.
- See: https://tailwindcss.com/docs/animation

Also, this is the first time we've used a *_url named route. I use these
so infrequently I had actually totally forgotten that we must specify
the host with `default_url_options` somewhere in the config. I'm not
really sure what the best approach for this is, but I went with
specifying it in the routes file. Other options would include putting
something similar into application.rb or into the environment-specific
configs directly.
- See: https://stackoverflow.com/questions/5691727/how-to-set-config-action-controller-default-url-options-host-on-per