Simplify welcome banner stream updates
We don't need to render a second/wrapper partial to get Welcome Banner
updates done. So we can just reduce this down to the layout partial and
be done with it.

Also, collapse the container/css code into the `turbo_frame_tag`. I've
decided that this has to be OK to do because, like it or not, this tag
does create an HTML element in the DOM. This can get in the way of
spacing rules, flex containers/contents, etc. So. We have to be able to
treat it as such. Which means we should feel free to add CSS classes to
this element as well--for the above-mentioned situations and more.