Add breadcrumbs (support) to page nav rendering
It's useful to know the context of the left/right arrows in page nav.
Especially for distinguishing between past Games nav vs past Games nav
in the context of a User. (The former becomes `Sweep Ops Archive >` nav
and the latter comes `<User Name> >` nav.)

To support this, simplify and flatten out styling for page nav in
general, by:
- Pushing styling up the stack to the application/_nav partial
- Add additional margin below the page nav element, but doing so
  dynamically based on the presence of the page_nav (so we don't have
  to remember to do it--and consistently update it as the design
  changes--every time we render a _page_nav partial)
- Making `turbo_frame_tag` elements do their own styling vs a secondary
  wrapper div doing the same