Move new Game views/models to games/new/*
This is in line with my previous restructuring of views/models. Now we
have major directory structure:
- games/new/
- games/current/
- games/just_ended/
- games/past/

And the outlier:
- custom_games/

Also, rename `find_or_create_game` -> `find_or_create_current_game`. I
wrestled with this one back in the day, but I think this is a better
indicator of actual function.