Move Random/Custom games under Games::New / "games/new"
This is in line with my previous restructuring of views/models. Now we
have major directory structure:
- games/new/
  - games/new/custom/   <- Instead of `custom_games/`
- games/current/
- games/just_ended/
- games/past/

The reason I didn't do this before is because I feel like it's
particularly ugly to have to read "Games::New::Custom::New.new", but I'm
now seeing the common directory structure for views as the most
important thing. Basically, having all the different Game views (which
is the hardest thing to keep track of logically) all sitting next to
each other is key.