Refactor game-state-view containers organization
This is a follow-up on my original restructuring (back in
cf2133751d7094ac02e2b1cc75e58feb0d3686a4) after playing with it for a
bit. I was having pain with using `show` pages as both a template and
a "partial".

e.g. having to do `<%# locals: (view: @view) %>` in the
games/new.html.erb template was hard to reason about. Templates
shouldn't have local variables, and especially not *sometimes*. This
just adds to cognitive load.

So now, we have a convention of using a "container" (partial + view
model) instead.