Move breadcrumb names into the Nav view model
If we need more complex breadcrumbs in the future, I think we can/should
accomplish this with View Models, and not with a more complex view
template. Thus, it's silly to use a partial layout.. and it adds some
complexity over normal expectations as is today, to boot.
- I also found out there are no current "conditional rendering" concerns
  for breadcrumb names, so removed this from the view template
  implementation (was previously implemented as `block_given?`)

Also, change the "Sweep Ops Archive" breadcrumb to just "Archive" for
simplicity. It also matches up with the page name in the Mobile View,
fwiw.