Add dynamic 400/404/406/422/500 error pages
Just for something a bit friendlier and a bit more on-theme than the
standard error pages that come with Rails.

Of note: the 404 page helps point out if perhaps a typo was made in the
URL.

These views are copied over from another project of mine, but also
customized to fit this project better.

Also, links to each of these views is available in the UI Portal under a
new "Error Pages" page.