Make _dev_flags overlay into a Portal toggle link
Can now click on the dev flags overlay at bottom-right of page (in
development env only) to toggle back and forth between the main portal
and the UI Portal.

To support this, we add a UIPortal::Layout with a specialized `portal`
String Inquirer that can easily answer the question for us. e.g.

```
layout.portal.ui?  # => `true` if in a UIPOrtal action
layout.portal.application?  # => `true` if in a base application action
```

This was also an opportunity to tune the UIPortal::Layout to fit just
what we need for that portal, vs using all of Application::Layout, like
we were previously.

Also, remove the old link to the UI Portal which was hidden as the
heart emoji in the footer text.