Abstract out Mine Density Validation for Custom Games
Reek prompted me to take this action, with its warning about
Board::Settings having too many methods. This is definitely a good chunk
of logic that is related to Board::Settings, but isn't core to it. i.e.
SRP was being violated before. This also just helps reduce the class
size + cognitive load of opening this file, so I'm in.

Also:
- Reorganize Board::Settings' methods a bit into a more sensible order
  (for me, as of now :P)
- Add a couple of missing tests to Board::SettingsTest