Fix up / disable remaining reek warnings
There isn't much else I want to do with the rest of the reek warnings,
so I'm just disabling them at this point. Cell has a lot of methods,
yes, but it kind of needs to. These are a lot of convenience methods
that help distill the game logic quite a bit. However, I did find one
method that was unused (still?): Cell#correctly_flagged? so remove it.
This puts the total methods to 20, and I think reek starts warning at
> 15.