Rework game states structure a bit more
I previously leaned towards adding CSS classes for spacing on
turbo_frame_tags, but the problem with that is that the original frame
tag then affects the content that it will be replaced with in the
future. So it seems that content should be up to its own styling, in all
cases. Perhaps the remaining reason to add CSS classes to turbo frames,
then, would be just to make that element, itself, behave properly / as
needed within the view. e.g. giving it a `block` class so that it
responds properly to a spacing class on its container element.
- Fixing this fixes spacing issues on Metrics - Games

Also, I previously passed up breaking out _container partials for past
User - Games and Metrics - Games. This was an oversight caused by these
views being further tucked away (out of sight out of mind). So this
commit makes structure consistent for User Games and for Metrics.

Also, fix the Game Show link from Metrics - Games to link back to the
`/metrics/games/<id>` path. i.e. we keep context on the source of the
Game Show link via Metrics pages, the same as for User Games.
- To support this, we expand the nav link for "Metrics" to make itself
  active when displaying Metrics::GamesController actions as well.