Add GameTrasactions (Create/Start/End)
This allows us to track which User:
- created a Game
- started a Game,
- made the final winning/losing move to end a Game.

To support this, we now have to pass a User object into the methods that
create, start, and end Games.

Note: For now, I'm keeping the Game#started_at and Game#ended_at
attributes. Though, I may remove these soon since they're redundant at
this point.

Also, fix an inconsistency in the test fixtures where the Game with
Status "Standing By" had a `started_at` value.

Also, fix ConsoleObjectBehaviors mix-in to not try to include has_many
associations as automatic Console objects. Because this gets in the way
of operations performed on the collection, such as `clear`.