Extract object inspection from *::Console objects
Moving object inspection into *::Console objects was an experiment in
processing optimization. But, I think it was a failed one. It just made
it too hard to see the pretty inspection results, especially when
chaining methods together in e.g. ARel queries.

Also, the Console object accessor / insulation layer was always a bit of
a doubling up on processing optimization. ObjectInspector itself has
scopes for dealing with this sort of thing, wherever necessary. For this
project, however, I can't think of any relevant cases (yet).