Fix User-specific Games nav
In the previous commit, Games::Past::Nav changed in a way that made it
clear that Users::Games::Nav really isn't a Games::Past::Nav. They were
code sharing, but the overlap was actually pretty small to start with.

This change also caused an error when trying to view User-specific
Games. So, to fix, I'm separating the 2 classes. This causes a bit of
code duplication, but not much.

Also, Games::Listings::Listing now expects a `context` while
Users::Games::Listings::Listing doesn't need one. This is problematic
since the latter is a subclass of the former. To fix this, we can make
the parent class only conditionally require a `context` and just have it
do the right thing if there isn't one. This works for me since the
context is also only conditionally needed by the main
Games::Listings::Listing -- when a Type filter is active. When on Type
filter is active we're just slicing out an empty Hash, which is the
same as defaulting `context` to nil and only pulling out the `type`
param/hash when it's not.
- I think the overall change here actually makes things easier as it
  changes the Games::Index.new param from `type_filter` (which isn't
  obvious what that means) to `context` (which is both obvious what it
  means and also allows for easier future expansion by being more
  generic).