Determine timer updates from starting timestamp instead
Previously, Game timer updates were based solely on incrementing the
number of elapsed seconds each second. But this was prone to lag in the
JS timer function, and could easily throw off the accuracy of the Game
timer over time (the longer the elapsed time, the greater the
disparity).

Instead, we can convert the starting "elapsed seconds" value to a
timestamp and then just always update the Game timer based on a
difference calculation from the starting timestamp. This way, regardless
of how delayed (or early) the JS interval function is, we'll still be
outputting the correct time, down to the second. Thus, increasing
accuracy.

Also, extract TimeParser out to a library file called ParseTime. This
code didn't otherwise change.