Setup kamal for deploys to my local network machine
STILL TO COME:
- Buy/use a Domain Name
- Consider VM hardware/sizing
- Verify Solid Queue (use broadcast_later*), and switch if it works
- Verify Solid Cache
- Verify Solid Cable

10.211.55.5 is the IP address for the Debian linux Virtual Machine
that's running on my Mac Mini on my local network. An SSH proxy (jump)
is configured via my Mac Mini to access this, which kamal uses for
deploys, etc.

NOTE: We'll need to add back the `proxy` setting to config/deploy.yml
once we actually get a domain name established.

There were many days of hacking on problems to get this to work. The
final hurdle, which I didn't see any help from anyone anywhere on, was
getting the database password to be recognized by the web server(s).
However, this ended up being as easy as just referencing
POSTGRES_PASSWORD in the env.secrets hash in config/deploy.yml.

Beyond this, I'm using 1Password for secrets storage, which requires
1Password to be open at deploy time. And Docker Hub as the docker image
registry, which requires Docker Desktop to be running at deploy time as
well.

Resources:
- https://kamal-deploy.org/docs/installation/
- https://www.driftingruby.com/episodes/kamal-2?query=kamal
- https://www.driftingruby.com/episodes/kamal-2-with-postgresql?query=kamal