Use Turbo to disconnect from :current_game stream on Game-end
Instead of using Stimulus JS to remove / disconnect from the
:current_game Turbo Stream, we can use Turbo to do the same by just
inserting a `turbo_stream.remove` call into the just-ended Game
container partial. i.e. when this partial is rendered Turbo will detect
the `remove` directive on the :current_game stream (removing it), which,
in turn, disconnects the user from the Stream.

This is easier and, I think more idiomatic for Turbo... still figuring
this out, though. It does, at least, remove the Stimulus dependency
which feels like a win.